# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starfront collaboration server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent/hello": {
      "post": {
        "summary": "Hello",
        "description": "Register a heartbeat and say what this rig is.\n\nThe profile is what makes the pre-flight check possible \u2014 telling somebody\ntheir 7 nm Ha will be rejected before they spend a night on it rather than\nafter.",
        "operationId": "hello_api_v1_agent_hello_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HelloRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelloResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/report": {
      "post": {
        "summary": "Report",
        "description": "Hand back what was actually captured, and hear whether it counts.\n\nJudged here rather than taken on trust, because the rules belong to the\nproject and an agent should not be able to mark its own homework. The\nverdict is advisory \u2014 a coordinator can overrule it.",
        "operationId": "report_api_v1_agent_report_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "post": {
        "summary": "Create Agent",
        "description": "Enrol a telescope and mint its token.\n\nAnybody signed in may enrol their own telescopes - that is what signing in\nis for - and the telescope is theirs from then on. The token is shown once,\nhere, and kept nowhere but the observatory it is for.",
        "operationId": "create_agent_api_v1_agents_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Agents",
        "description": "The owner sees every telescope; a member sees their own.",
        "operationId": "list_agents_api_v1_agents_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "post": {
        "summary": "Create Project",
        "operationId": "create_project_api_v1_projects_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Projects",
        "description": "Open to anyone: browsing is how people find a collab to join.",
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": ""
          },
          "created": {
            "type": "number",
            "title": "Created"
          },
          "seen": {
            "type": "number",
            "title": "Seen",
            "default": 0.0
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RigProfile"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Profile"
          },
          "presence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Presence"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presence"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name",
          "created"
        ],
        "title": "Agent",
        "description": "An enrolled telescope, without its token."
      },
      "AgentCreated": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Shown once. Keep it on the telescope it is for."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "agent",
          "token"
        ],
        "title": "AgentCreated"
      },
      "AgentList": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/Agent"
            },
            "type": "array",
            "title": "Agents"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "agents"
        ],
        "title": "AgentList"
      },
      "AgentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "owner": {
            "type": "string",
            "maxLength": 80,
            "title": "Owner",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AgentRequest"
      },
...
      "Contribution": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "default": ""
          },
          "task": {
            "type": "string",
            "title": "Task",
            "default": ""
          },
          "night": {
            "type": "string",
            "title": "Night",
            "default": ""
          },
          "filterName": {
            "type": "string",
            "title": "Filtername",
            "default": ""
          },
          "panel": {
            "type": "string",
            "title": "Panel",
            "default": ""
          },
          "frames": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames",
            "default": 0
          },
          "seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds",
            "default": 0.0
          },
          "exposure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure",
            "description": "Sub length, seconds.",
            "default": 0.0
          },
          "footprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region"
              },
              {
                "type": "null"
              }
            ],
            "description": "The solved footprint of the panel."
          },
          "scale": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale",
            "description": "Arcseconds per pixel."
          },
          "focalLength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focallength"
          },
          "hfr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hfr",
            "description": "Mean star size over the night, arcseconds."
          },
          "guideRms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guiderms",
            "description": "Mean guiding error, arcseconds."
          },
          "moonIllumination": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moonillumination"
          },
          "moonSeparation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moonseparation",
            "description": "Degrees."
          },
          "calibrated": {
            "type": "boolean",
            "title": "Calibrated",
            "default": false
          },
          "bandpass": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bandpass",
            "description": "nm, of the filter used."
          },
          "colour": {
            "type": "boolean",
            "title": "Colour",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "Contribution",
        "description": "One night's work on one filter and panel, as measured."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HelloRequest": {
        "properties": {
          "protocol": {
            "type": "integer",
            "title": "Protocol",
            "default": 1
          },
          "profile": {
            "$ref": "#/components/schemas/RigProfile"
          },
          "presence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Presence"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "HelloRequest",
        "description": "A telescope saying what it is and, if it likes, where it points."
      },
      "HelloResponse": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "protocol": {
            "type": "integer",
            "title": "Protocol"
          },
          "serverTime": {
            "type": "number",
            "title": "Servertime"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "agent",
          "name",
          "protocol",
          "serverTime"
        ],
        "title": "HelloResponse"
      },
...
      "Presence": {
        "properties": {
          "ra": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ra",
            "description": "Hours."
          },
          "dec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dec",
            "description": "Degrees."
          },
          "state": {
            "type": "string",
            "title": "State",
            "default": ""
          },
          "target": {
            "type": "string",
            "title": "Target",
            "default": ""
          },
          "project": {
            "type": "string",
            "title": "Project",
            "default": ""
          },
          "telescope": {
            "type": "string",
            "title": "Telescope",
            "description": "What to call the telescope on the group's chart, such as its equipment profile's name.",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "Presence",
        "description": "Where a telescope points and what it is doing, as it chooses to say."
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "coordinator": {
            "type": "string",
            "title": "Coordinator",
            "default": ""
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": ""
          },
          "created": {
            "type": "number",
            "title": "Created"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "closed"
            ],
            "title": "Status",
            "default": "open"
          },
          "payload": {
            "$ref": "#/components/schemas/ProjectPayload"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name",
          "created",
          "payload"
        ],
        "title": "Project"
      },
...
      "ProjectEnvelope": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Project"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectEnvelope"
      },
      "ProjectList": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "projects"
        ],
        "title": "ProjectList"
      },
      "ProjectPayload": {
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "kind": {
            "type": "string",
            "enum": [
              "single",
              "mosaic"
            ],
            "title": "Kind",
            "default": "mosaic"
          },
          "requirements": {
            "$ref": "#/components/schemas/Requirements"
          },
          "goals": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Goals",
            "description": "Depth wanted at every point, per filter, in hours."
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "region"
        ],
        "title": "ProjectPayload",
        "description": "A project's definition, as stored."
      },
      "ProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "coordinator": {
            "type": "string",
            "maxLength": 80,
            "title": "Coordinator",
            "default": ""
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "kind": {
            "type": "string",
            "enum": [
              "single",
              "mosaic"
            ],
            "title": "Kind",
            "default": "mosaic"
          },
          "requirements": {
            "$ref": "#/components/schemas/Requirements"
          },
          "goals": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Goals",
            "description": "Hours per filter at every point."
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "region"
        ],
        "title": "ProjectRequest"
      },
...
      "Recorded": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "accepted",
          "duplicate",
          "verdict"
        ],
        "title": "Recorded"
      },
      "Region": {
        "properties": {
          "ra": {
            "type": "number",
            "title": "Ra",
            "description": "Right ascension of the centre, degrees."
          },
          "dec": {
            "type": "number",
            "title": "Dec",
            "description": "Declination of the centre, degrees."
          },
          "width": {
            "type": "number",
            "title": "Width",
            "description": "Width on the sky, degrees."
          },
          "height": {
            "type": "number",
            "title": "Height",
            "description": "Height on the sky, degrees."
          },
          "rotation": {
            "type": "number",
            "title": "Rotation",
            "description": "Position angle to shoot at, degrees.",
            "default": 0.0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ra",
          "dec",
          "width",
          "height"
        ],
        "title": "Region",
        "description": "A rectangle of sky. Degrees throughout; `ra` is degrees, not hours.\n\n`width` and `height` are angles on the sky. `rotation` is the position\nangle contributing cameras should shoot at; the rectangle itself is\nnorth-up."
      },
      "ReportRequest": {
        "properties": {
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/Contribution"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Contributions"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ReportRequest",
        "description": "One night's work: one record per filter and panel."
      },
      "ReportResponse": {
        "properties": {
          "recorded": {
            "items": {
              "$ref": "#/components/schemas/Recorded"
            },
            "type": "array",
            "title": "Recorded"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "recorded"
        ],
        "title": "ReportResponse"
      },
...
      "RigProfile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "focalLength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focallength",
            "description": "Millimetres."
          },
          "pixelSize": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pixelsize",
            "description": "Microns, unbinned."
          },
          "sensorWidth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensorwidth",
            "description": "Pixels, unbinned."
          },
          "sensorHeight": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensorheight",
            "description": "Pixels, unbinned."
          },
          "binning": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binning",
            "default": 1
          },
          "filters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Filters",
            "description": "Filter name to bandpass in nm, or null if unknown."
          },
          "colour": {
            "type": "boolean",
            "title": "Colour",
            "description": "A one-shot colour camera.",
            "default": false
          },
          "rotation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotation",
            "description": "Camera position angle in degrees when it cannot be turned; null when a rotator can set any angle."
          },
          "typicalHfr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typicalhfr",
            "description": "Arcseconds."
          },
          "typicalGuideRms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typicalguiderms",
            "description": "Arcseconds."
          },
          "exposures": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Exposures",
            "description": "Sub length in seconds per filter: what the rig's darks are built for."
          },
          "hoursPerNight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hourspernight",
            "description": "Hours a night the rig gives; null means as long as targets are up."
          },
          "windowFrom": {
            "type": "string",
            "title": "Windowfrom",
            "description": "Local clock time, such as \"21:00\".",
            "default": ""
          },
          "windowTo": {
            "type": "string",
            "title": "Windowto",
            "description": "Local clock time.",
            "default": ""
          },
          "scale": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale",
            "description": "Arcseconds per pixel, computed."
          },
          "field": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Field of view [width, height] in degrees, computed."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RigProfile",
        "description": "What a telescope is and what it achieves, sent with every hello."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "unverified": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unverified"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "overriddenBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overriddenby"
          },
          "overrideReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridereason"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "Verdict"
      },
...
    }
  }
}